don't waste space in registry for implicit final state
there is an implicit final state which is final, has zero
transitions, and no final output value.  this is very common
and recorded in file with special addr of 0.  however, we were
wasting a slot in our cache to remember this information which
signficantly hurt file compatibility with fst crate.

in the future additional work is required to clean up the builder
as we really only need to remember the 0, but for now we
keep yet another reference to the state.