change automaton opaque state to int
previously we used interface{} which is more flexible
but, also requires type assertions through all use of
the automaton, because you never know if the state passed in
is one you can work with.

the change is to now use int to represent the state
the int is still opaque (0 or negative might mean different
things to different impls) but int still seems reasonable.