added a Reader abstraction for optimized Get()'s
Some applications may need to perform repeated Get() lookups on an
FTS, in which case it may be beneficial for the app to create one or
more Reader instances, where each Reader instance may only be used by
a single thread at a time.

The underlying optimization of a Reader instance is to avoid
allocations of ftsStateV1 instances.

Change-Id: I0eec3aa5956db914b876d065090a13722174b633
Reviewed-on: http://review.couchbase.org/92461
Reviewed-by: Marty Schoch <marty.schoch@gmail.com>
Tested-by: Steve Yen <steve.yen@gmail.com>