optimize FSTIterator.next() with tigher inner loop
Previously, when there was only a single loop (the 'OUTER' loop) in
FSTIterator.next(), when the curr and autCurr did not match, the code
would bump the lastOffset, go back to the top of the outer loop, and
(wastefully) reinitialize and re-check the curr and autCurr states
even though they hadn't changed.

This optimization introduces an inner loop which allows the nextOffset
to be more efficiently moved fowards when the curr and autCurr don't
match.

Change-Id: Ie6724cdc87a3749933a7a48216e5a3ccd8f9b676
Reviewed-on: http://review.couchbase.org/96970
Reviewed-by: Abhinav Dangeti <abhinav@couchbase.com>
Tested-by: Steve Yen <steve.yen@gmail.com>