fix bug in iterator next optimization
The iterator next method has an optimization, when no more
transitions are possible, and the iterator must backtrack,
it attempts to go back through all single transition states
in one shot.  For example, the current key had some long
unique suffix, we need to back track through that entire
portion.  The optimization attempts to identify this and
handle it faster, rather than going back one state at a time
which involes more checks, and is therefore slower.

However, there is an edge case this optimzation does not handle
correctly.  If the entire stack has one transition, we can go
all the way back to the root, and the code was not written to
address this correctly.  It somewhat safely still succeeds in
going back one state at a time, but it keeps attempting the
optimization on each iteration, making it very slow.

This fix identifies the case where all examined states satisfy
the check, and now sets the value correctly to apply the
optimization.

The included test has no assertions, but does timeout after 10m
on my machine.  With the fix, the test completes in 0.5s

Change-Id: I9744ae012a4e2e95e01d1a709905b48dc693adbd
Reviewed-on: http://review.couchbase.org/c/vellum/+/134677
Well-Formed: Build Bot <build@couchbase.com>
Tested-by: Marty Schoch <marty.schoch@gmail.com>
Tested-by: Abhinav Dangeti <abhinav@couchbase.com>
Reviewed-by: Sreekanth Sivasankaran <sreekanth.sivasankaran@couchbase.com>
Reviewed-by: Abhinav Dangeti <abhinav@couchbase.com>